+++ podpointclient/endpoints.py
@@ -7,8 +7,9 @@ PODS = '/pods'
 UNITS = '/units'
 CHARGE_SCHEDULES = '/charge-schedules'
 CHARGES = '/charges'
+CHARGE_OVERRIDE = '/charge-override'
 FIRMWARE = '/firmware'
 
 API_BASE = 'api.pod-point.com'
-API_VERSION = 'v4'
+API_VERSION = 'v5'
 API_BASE_URL = 'https://' + API_BASE + '/' + API_VERSION