+++ podpointclient/pod.py
@@ -9,6 +9,8 @@ from strenum import StrEnum, KebabCaseStrEnum
 from .helpers.functions import lazy_convert_to_datetime, lazy_iso_format_datetime
 from .schedule import Schedule, ScheduleStatus
 from .charge import Charge
+from .charge_mode import ChargeMode
+from .charge_override import ChargeOverride
 
 
 class StatusName(StrEnum):
@@ -242,6 +244,11 @@ class Pod:
                 )
             )
 
+        self.charge_override = None
+        charge_override_data = data.get('charge_override', None)
+        if charge_override_data is not None:
+            self.charge_override = ChargeOverride(data=charge_override_data)
+
 
     @property
     def dict(self) -> Dict[str, Any]:
@@ -272,7 +279,8 @@ class Pod:
             "total_charge_seconds": self.total_charge_seconds,
             "current_kwh": self.current_kwh,
             "total_cost": self.total_cost,
-            "firmware": None
+            "firmware": None,
+            "charge_override": None
         }
 
         for status in self.statuses:
@@ -289,12 +297,32 @@ class Pod:
         if isinstance(self.firmware, Firmware):
             dictionary['firmware'] = self.firmware.dict
 
+        if isinstance(self.charge_override, ChargeOverride):
+            dictionary['charge_override'] = self.charge_override.dict
+
         return dictionary
 
     def to_json(self) -> str:
         """JSON representation of a Pod"""
         return json.dumps(self.dict, ensure_ascii=False)
+    
+    @property
+    def charge_mode(self) -> ChargeMode:
+        """what is the current charge override mode?"""
+        override = self.charge_override
 
+        if override is None or override.ppid is None:
+            return ChargeMode.SMART
+
+        if override.active:
+            return ChargeMode.OVERRIDE
+
+        elif override.requested_at is not None and self.recieved_at is not None and self.ends_at is None:
+            return ChargeMode.MANUAL
+
+        else:
+            _LOGGER.warn("Unable to caclculate charge mode")
+            return None
 
     @dataclass
     class Model: