+++ tests/test_api_wrapper.py
@@ -39,6 +39,17 @@ async def test_put(aiohttp_client):
         assert "OK" == await result.text()
 
 @pytest.mark.asyncio
+async def test_delete(aiohttp_client):
+  with aioresponses() as m:
+    m.put('https://google.com/api/v1/test?foo=bar', status=204, body="OK")
+
+    async with aiohttp.ClientSession() as session:
+      wrapper = APIWrapper(session)
+      async with await wrapper.delete("https://google.com/api/v1/test", body={}, headers={}, params={"foo": "bar"}) as result:
+        assert 204 == result.status
+        assert "OK" == await result.text()
+
+@pytest.mark.asyncio
 async def test_401(aiohttp_client):
   with aioresponses() as m:
     m.get('https://google.com/api/v1/test?foo=bar', status=401, body="AuthError")